refactor(form): script setup
