+++ src/packages/__VUE/form/types.ts
@@ -1,6 +1,18 @@
-import { FormItemRule, FormItemLabelPosition, FormItemStarPosition } from '../formitem/types';
-
 export const FORM_KEY = Symbol('nut-form');
+
+export interface FormItemRuleWithoutValidator {
+  [key: string]: any;
+  regex?: RegExp;
+  required?: boolean;
+  message?: string;
+}
+
+export interface FormItemRule extends FormItemRuleWithoutValidator {
+  validator?: (
+    value: any,
+    ruleCfg: FormItemRuleWithoutValidator
+  ) => boolean | Promise<string> | Promise<boolean> | Promise<void> | Promise<unknown>;
+}
 
 export type FormRules = {
   [key: string]: FormItemRule[];
@@ -11,10 +23,10 @@ export type FormRule = {
   rules: FormItemRule[];
 };
 
-export type ErrorMessage = {
+export type FormErrorMessage = {
   prop: string;
   message: string;
 };
 
-export type FormLabelPosition = FormItemLabelPosition;
-export type FormStarPosition = FormItemStarPosition;
+export type FormLabelPosition = 'left' | 'right' | 'top';
+export type FormStarPosition = 'left' | 'right';
